+++ CHANGELOG.md
@@ -26,6 +26,8 @@
   [#47](https://github.com/oddbird/sassdoc-theme-herman/issues/47)
 - BREAKING: Replace `minifiedIcons` option with `customHTML` option
   (only included in rendered `@example` annotations).
+- BREAKING: Nest `templatepath` and `nunjucksEnv` options under new `nunjucks`
+  options object: `nunjucks.templatepath` and `nunjucks.environment`.
 
 
 ## 1.0.0-alpha.15: 2017-10-30